test(cli): consolidate root help fast-path coverage
